fix: improve code accorging to tips
