+++ src/pages/Home/Home.js
@@ -1,11 +1,24 @@
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
 import Slider from 'components/Slider';
 import IconsSection from 'components/IconsSection';
 import ShowProductsSection from 'components/ShowProductsSection';
 import BlogLayout from 'components/BlogSection/BlogLayout';
 import useTilg from 'tilg';
+import { fetchProducts } from 'store/products';
+import { fetchCategories } from 'store/categories';
 
 const Home = () => {
   useTilg();
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(fetchCategories());
+  }, [dispatch]);
+
+  useEffect(() => {
+    dispatch(fetchProducts());
+  }, [dispatch]);
 
   return (
     <>
