+++ src/stylesheets/base/_variables.sass
@@ -14,6 +14,31 @@ $error-color: #9b3535
 $mx-xl: 15%
 $mx-l: 10%
 $mx-s: 5%
-$m-base: 24px
-$m-l: calc($m-base / 2) // 12px
-$m-s: calc($m-base / 3) //8px
+$marg-base: 10px
+$marg-2: calc($marg-base * 1.5) // 15px
+$marg-3: calc($marg-base * 2) // 20px
+$marg-4: calc($marg-base * 3) // 30px
+$marg-5: calc($marg-base * 4) // 40px
+$marg-6: calc($marg-base * 5) // 50px
+$marg-7: calc($marg-base * 10) // 100px
+$marg-8: calc($marg-base * 15) // 150px
+
+
+
+//---- font-sizes -------------------------------------------
+$l-3: 72px
+$l-2: 64px
+$l-1: 50px
+$b-4: 48px
+$b-3: 38px
+$b-2: 36px
+$b-1: 32px
+$m-5: 24px
+$m-4: 22px
+$m-3: 20px
+$m-2: 18px
+$m-1: 16px
+$s-4: 15px
+$s-3: 14px
+$s-2: 12px
+$s-1: 10px
