+++ src/stylesheets/components/BlogSection/BlogLayout.sass
@@ -1,4 +1,4 @@
 .blog-layout
   @include flex($direction: column)
   @include alignX
-  margin-bottom: 50px
+  margin-bottom: $marg-6
