+++ src/stylesheets/components/BlogSection/BlogPost.sass
@@ -21,7 +21,7 @@
       background-color: #F4F5F8
 
       &__title
-          font-size: 20px
+          font-size: $m-3
           font-family: Raleway
           display: -webkit-box
           -webkit-line-clamp: 2
@@ -30,15 +30,15 @@
 
       &__info
           @include flex($main: flex-start)
-          font-size: 10px
-          margin: 10px 0
+          font-size: $s-1
+          margin: $marg-base 0
 
           &__author
               margin-right: 5px
 
       &__text
-          font-size: 14px
-          margin-bottom: -20px
+          font-size: $s-3
+          margin-bottom: -$marg-base
           display: -webkit-box
           -webkit-line-clamp: 4
           -webkit-box-orient: vertical
