+++ src/stylesheets/components/Cart/CartCounter.sass
@@ -1,6 +1,6 @@
 .counter
   @include flex
-  margin-right: 10px
+  margin-right: $marg-2
 
   &__left,
   &__right
@@ -8,7 +8,7 @@
     background-color: transparent
     height: 50px
     width: 25px
-    font-size: 20px
+    font-size: $m-3
     text-align: center
     transition: color 0.3s ease 0s
