+++ src/stylesheets/components/CartButton/CartButton.sass
@@ -13,18 +13,18 @@
 
   &__icon
     color: $accent-color
-    font-size: 20
+    font-size: $m-4
 
   &__items
     color: $white-color
-    font-size: 14px
+    font-size: $s-3
     font-weight: bold
     padding: 8px
 
   &__amount
     color: $accent-color
     background-color: $white-color
-    font-size: 14px
+    font-size: $s-3
     font-weight: bold
     width: 100%
     text-align: center
