+++ src/stylesheets/components/CartModal/CartModal.sass
@@ -14,15 +14,15 @@
     border-bottom: 1px solid #ccc
 
     &__title
-      padding: 20px
-      font-size: 20px
+      padding: $marg-3
+      font-size: $m-3
 
     &__close
-      margin-right: 20px
+      margin-right: $marg-3
       @include closeBtn
 
   &__items
-        font-size: 20px
+    font-size: $m-3
 
   &__controls
       @include flex
@@ -31,13 +31,13 @@
 
       &__btn
         width: 100%
-        margin: 20px
+        margin: $marg-3
         border: none
         border-radius: 6px
         background-color: $accent-color
         color: $white-color
-        font-size: 16px
-        padding: 15px
+        font-size: $m-1
+        padding: $marg-2
         text-align: left
         cursor: pointer
