+++ src/stylesheets/components/IconsSection/IconItem.sass
@@ -3,17 +3,17 @@
   width: 180px
   height: 180px
   border-radius: 50%
-  margin: 0 15px
+  margin: 0 $marg-2
 
   &__icon
     @include flex
-    font-size: 4rem
+    font-size: $l-2
     color: $font-less-dark
     position: relative
     width: 100%
     height: 100%
     border-radius: 50%
-    margin: 0 15px
+    margin: 0 $marg-2
 
     &__border
       border: 2px solid $font-less-dark
@@ -35,10 +35,10 @@
       @include flex($direction: column)
                 
       #{$root}__icon__text
-        font-size: 1rem
+        font-size: $m-1
         font-family: Raleway
         margin: 12px 0 0 0
 
 @media (max-width: $xl)
   .icon-section
-    margin: 15px
+    margin: $marg-2
