+++ src/stylesheets/components/IconsSection/IconsSection.sass
@@ -1,5 +1,5 @@
 .icons-section
   width: 100%
-  padding: 100px 0
+  padding: $marg-7 0
   @include flex
   flex-wrap: wrap
