+++ src/stylesheets/components/Products/GridItem.sass
@@ -1,7 +1,7 @@
 .gridItem
   $root: &
   @include flex($direction: column)
-  margin: 0 0 30px 0
+  margin: 0 0 $marg-4 0
   cursor: pointer
   position: relative
 
@@ -26,7 +26,7 @@
 
     &__text
       font-family: Montserrat
-      font-size: 16px
+      font-size: $m-1
       line-height: 24px
       font-weight: 500
