+++ src/stylesheets/components/Search/Search.sass
@@ -11,7 +11,7 @@
   &__controls
     @include flex
     width: 40%
-    padding: 30px 0 
+    padding: $marg-4 0 
 
     &__input
       width: 100%
@@ -20,7 +20,7 @@
       border: 1px solid #ccc
       border-top-left-radius: 5px
       border-bottom-left-radius: 5px
-      font-size: 16px
+      font-size: $m-1
 
     &__button
       width: 70px
@@ -30,10 +30,10 @@
       border-bottom-right-radius: 5px
       background-color: $accent-color
       color: $white-color
-      font-size: 20px
+      font-size: $m-3
       cursor: pointer
 
   &__close
-    margin-bottom: 20px
+    margin-bottom: $marg-3
     @include closeBtn
