+++ src/stylesheets/components/Slider/RoundButton.sass
@@ -7,7 +7,7 @@
   cursor: pointer
 
   &__arrow
-    font-size: 2rem
+    font-size: $b-1
     line-height: 40px
 
   &:hover
