+++ src/stylesheets/components/Slider/SliderContent.sass
@@ -10,20 +10,20 @@
     margin: 0
 
     &__subtitle
-     font-size: 24px
+     font-size: $m-5
      color: $font-less-dark 
       font-family: Raleway
       animation: fadeIn 0.8s ease 
 
     &__title
-      font-size: 72px
+      font-size: $l-3
       color: $font-dark 
       font-family: Raleway
-      margin: 10px 0
+      margin: $marg-base 0
       animation: fadeIn 1.2s ease 
 
     &__description
-      margin-right: 100px
+      margin-right: $marg-7
       color: $font-less-dark
       animation: fadeIn 1.6s ease 
 
@@ -35,7 +35,7 @@
 @media (max-width: $xl)
   .slider-content__element
     &__title
-      font-size: 50px
+      font-size: $l-1
     &__image
       width: 200px
 
@@ -49,15 +49,15 @@
       &__description
         margin-right: 0
       &__image
-        margin-bottom: 50px
+        margin-bottom: $marg-6
         max-height: 180px
 
 @media (max-width: $s)
   .slider-content
       &__element
         &__subtitle
-          font-size: 20px
+          font-size: $m-3
         &__title
-          font-size: 36px
+          font-size: $b-2
         &__image
-          margin-bottom: 20px
+          margin-bottom: $marg-3
