+++ src/stylesheets/components/UI/Button.sass
@@ -1,6 +1,6 @@
 .button 
   color: $white-color
-  font-size: 16px
+  font-size: $m-1
   width: 150px
   height: 50px
   font-family: Raleway
@@ -18,7 +18,7 @@
   &.-small
     width: 90px
     height: 35px
-    font-size: 10px
+    font-size: $s-1
 
   &.-color-reverse
     color: $white-color
