+++ src/stylesheets/components/UI/ButtonRound.sass
@@ -7,7 +7,7 @@
   position: relative
   height: 30px
   width: 30px
-  font-size: 15px
+  font-size: $s-4
 
   &:hover
     color: $white-color
@@ -29,7 +29,7 @@
       right: calc(100% + 4px)
       width: 80px
       display: none
-      font-size: 10px
+      font-size: $s-1
       top: 50%
       transform: translateY(-50%)
