+++ src/stylesheets/components/UI/ButtonSlim.sass
@@ -1,11 +1,11 @@
 .buttonSlim 
   color: $font-less-dark
-  font-size: 18px
+  font-size: $m-2
   font-family: Raleway
   background-color: transparent
   border: none
   cursor: pointer 
-  margin: 10px 
+  margin: $marg-base
 
   &--isActive
     color: $accent-color
