+++ src/stylesheets/components/UI/ButtonSlimContainer.sass
@@ -1,4 +1,4 @@
 .btn-container
   @include flex
-  margin: 30px 0
+  margin: $marg-4 0
   flex-wrap: wrap
