+++ src/stylesheets/components/UI/Error.sass
@@ -1,5 +1,5 @@
 .error
   color: $error-color
   border: 1px solid $error-color
-  padding: 10px 30px
-  margin: 50px 0
+  padding: $marg-base $marg-4
+  margin: $marg-6 0
