+++ src/stylesheets/components/UI/Footer.sass
@@ -1,7 +1,7 @@
 .footer 
   background-color: #2F333A
-  margin-top: 150px
-  padding: 100px $mx-xl
+  margin-top: $marg-8
+  padding: $marg-7 $mx-xl
   @include flex($main: space-between, $cross: flex-start)
     
 
@@ -15,14 +15,14 @@
 
     &__title
       font-family: Raleway
-      font-size: 14px
-      margin-bottom: 10px
+      font-size: $s-3
+      margin-bottom: $marg-base
 
     &__description
-      font-size: 12px
+      font-size: $s-2
 
     &__list
-      font-size: 12px
+      font-size: $s-2
 
       & li
         margin-bottom: 5px
@@ -39,7 +39,7 @@
       &__icon
         width: 20px
         height: 20px
-        margin: 10px 10px 0 0
+        margin: $marg-base $marg-base 0 0
         cursor: pointer
         border: 1px solid $white-color
         border-radius: 50%
