+++ src/stylesheets/components/UI/Header.sass
@@ -8,11 +8,11 @@
 
     &__icon
       color: $accent-color
-      font-size: 2rem
-      margin-right: $m-s
+      font-size: $b-1
+      margin-right: $marg-base
         
     &__text
-      font-size: 2rem
+      font-size: $b-1
       font-weight: bold
 
   &__top-menu 
@@ -20,10 +20,10 @@
 
       &__control
         position: relative
-        margin: 0 $m-base
+        margin: 0 24px
 
         &__icon 
-          font-size: 24px
+          font-size: $m-5
           color: $primary-color
           cursor: pointer
 
@@ -34,7 +34,7 @@
           background-color: $accent-color
           border-radius: 50%
           color: $white-color
-          font-size: 0.8rem
+          font-size: $s-2
           height: 17px
           line-height: 17px
           position: absolute
