+++ src/stylesheets/components/UI/LoadingSpinner.sass
@@ -1,7 +1,7 @@
 .small
   width: 25px
   height: 25px
-  margin: 10px
+  margin: $marg-base
   position: relative
 
   &__outer
