+++ src/stylesheets/components/UI/Navbar.sass
@@ -17,16 +17,16 @@
 
     &__item 
       color: $white-color
-      font-size: 14px
+      font-size: $s-3
       letter-spacing: 0.5px
-      margin: 0 $m-base
+      margin: 0 24px
       cursor: pointer
       position: relative
       line-height: 60px
 
       &__dropdown-content 
         display: none
-        font-size: 14px
+        font-size: $s-3
         position: absolute
         top: 60px
         left: 20px
@@ -46,7 +46,7 @@
         @include hover  
 
       &__arrow  
-        font-size: 10px
+        font-size: $s-1
         position: absolute
         top: 1px
         right: -8px
