+++ src/stylesheets/components/UI/Wrapper.sass
@@ -1,4 +1,4 @@
 .show-container 
   @include flex($direction: column)
   @include alignX
-  margin-bottom: 50px
+  margin-bottom: $marg-6
